Move the rest of tests not specific to PG from the sqlparser_postgres.rs
